Refactored chat-related components
