+++ src/app/app.routes.ts
@@ -1,10 +1,10 @@
 import { Routes } from '@angular/router';
-import { DashboardComponent } from './dashboard/dashboard.component';
+import { ChatComponent } from './chat/chat.component';
 
 export const routes: Routes = [
     {
         path: "",
-        component: DashboardComponent,
-        title: "Agri-Multimodel"
+        component: ChatComponent,
+        title: "AgrIA"
     }
 ];
