+++ src/app/chat/chat-assistant/chat-assistant.component.css
