+++ src/app/chat/chat.component.css
@@ -1,4 +1,4 @@
-#dashboard {
+#chat {
     display: flex;
     flex-direction: column;
     align-items: center;
@@ -142,7 +142,7 @@
   }
 
   @media (max-width: 1024px) {
-    #dashboard .top-section {
+    #chat .top-section {
       flex-direction: column;
       align-items: center;
       gap: 1.5rem;
@@ -158,7 +158,7 @@
   }
   
   @media (max-width: 768px) {
-    #dashboard {
+    #chat {
       margin: 1rem;
     }
