+++ src/app/chat/chat.component.html
@@ -1,4 +1,4 @@
-<div id="dashboard">
+<div id="chat">
     <div class="content">
         <div class="top-section">
             <div id="image">
